Moved dot11 parsing logic to standalone function. First step to allow for fuzzing of the parser.
