+++ tap/src/protocols/parsers/dot11/mod.rs
@@ -1,2 +1,4 @@
 pub mod management;
 pub mod data;
+pub mod dot11_header_parser;
+
